[inductor] don't allow triton config pre_hook
The caching autotuner caches triton configs, and it doesn't try to hash or save the pre_hook from the config if it exists. If we had a config that had a pre_hook, then we might autotune -> save the config (without the pre_config) -> later load the saved config and try to run it, but this time without the pre_hook.

So this PR adds an assert and deletes the pre_hook handling. We can be confident that we didn't have functional pre_hooks, because the pre_hook handling tries to use `self.arg_name`, which doesn't exist.

ghstack-source-id: d80dd5219b9460740a30e89f11bccbbfea9046f8
Pull Request resolved: https://github.com/pytorch/pytorch/pull/134633